fix: rename and simplify commands
