+++ pkg/cmd/context/status/status.go
