+++ pkg/cmd/context/view/view.go
